WIP build and deploy
