+++ wizard/src/components/WizardShell.tsx
@@ -1,7 +1,9 @@
-import { ReactNode } from 'react';
+import { ReactNode, useState, useCallback } from 'react';
 import { useWizard } from '../hooks/useWizardState';
 import { StepIndicator } from './StepIndicator';
 import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/24/solid';
+import { BookmarkIcon, CheckIcon } from '@heroicons/react/24/outline';
+import { buildShareUrl } from '../utils/shareUrl';
 
 interface WizardShellProps {
   children: ReactNode;
@@ -36,6 +38,7 @@ export function WizardShell({
   prevLabel = 'Back',
 }: WizardShellProps) {
   const { state, dispatch } = useWizard();
+  const [copied, setCopied] = useState(false);
 
   const handleNext = () => {
     if (onNext) {
@@ -52,6 +55,17 @@ export function WizardShell({
     dispatch({ type: 'PREV_STEP' });
   };
 
+  const copyShareUrl = useCallback(async () => {
+    const url = buildShareUrl(state.config);
+    try {
+      await navigator.clipboard.writeText(url);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    } catch {
+      setCopied(false);
+    }
+  }, [state.config]);
+
   const isFirstStep = state.currentStep === 0;
   const isLastStep = state.currentStep === STEP_NAMES.length - 1;
 
@@ -63,7 +77,20 @@ export function WizardShell({
           <p className="mt-3 text-lg text-white/70">Configure your Dismissible API with ease</p>
         </div>
 
-        <div className="card p-6 md:p-8">
+        <div className="card p-6 md:p-8 relative">
+          <button
+            type="button"
+            onClick={copyShareUrl}
+            title="Copy share URL to save progress"
+            className="absolute top-6 right-6 p-2 rounded-lg border border-white/10 bg-white/5 text-white/70 hover:border-white/20 hover:bg-white/10 hover:text-white transition-all duration-200 z-10"
+          >
+            {copied ? (
+              <CheckIcon className="w-5 h-5" />
+            ) : (
+              <BookmarkIcon className="w-5 h-5" />
+            )}
+          </button>
+
           <StepIndicator
             currentStep={state.currentStep}
             totalSteps={STEP_NAMES.length}
